Create a plugin for collecting usage statistics
The following data might be collected:
1) Builds count
2) Used build systems
3) Opened examples
4) Used kits (detailed info)
5) Modes usage
6) QML designer usage time
7) License type

Also all common data (system info, graphic, app version, etc.) might be
collected as well.

Plugin is built atop of KUserFeedback framework and uses its
architecture and routines for collecting and sending data. KUserFeedback
is not included to this repository and supposed to be built and
installed separately by CI. Static build of KUserFeedback is used, so
plugin doesn't depend on any 3rd-party dynamic libraries.

URL for sending data and authentication key should be set during the
build process.

Change-Id: Ibf0cf4986031689b5f46925664cebadcb0fe3aef