Support building with CMake
Adds a meta-build that builds & installs extra-cmake-modules
and kuserfeedback, and finally src. Note that the meta-build uses
ExternalProject_Add, which means that any CMake variables have
to be passed explicitly to the sub-projects.

The new git submodule extra-cmake-modules allow us to use an
unchanged version of kuserfeedback/CMakeLists.txt

A typical call might look like this:

  export CMAKE_PREFIX_PATH=$QTDIR;$CREATOR_BUILD_DR
  cmake -DUSP_AUTH_KEY="..." -DUSP_SERVER_URL="..." -DCMAKE_INSTALL_PREFIX=$CREATOR_INSTALL_DIR

Change-Id: I4cc7b6936cad119156f19474c57045296ff6704e
Reviewed-by: Vitaly Fanaskov <vitaly.fanaskov@qt.io>