telemetry: Make sure to depend on the QtCreator dependencies
The plugin is being built using ExternalProject_Add, which means
that it will use QtCreator targets from the exported QtCreator build.

The trigger of the build is done from a super build including
QtCreator. Adding the QtCreator dependencies makes sure that the
binaries referenced from the exported build are first created.

Change-Id: I0bc4ea85170bf4beed869ef2205828e7fbc9b1a8
Reviewed-by: Tim Jenssen <tim.jenssen@qt.io>