Add information about opened Qt examples
Check if the project that was opened (parsed) is in a subdirectory of
any Qt version's examples path.
If so, assume that it is a Qt example.
Log a hash of the relative path below the Qt example path,
the full hash ID of the project path,
and the Qt version number.

Fixes: QTCREATORBUG-32879
Change-Id: I41560b23145f2b0a8b8e5905b907bd10df54a2ec
Reviewed-by: Kai Köhne <kai.koehne@qt.io>